requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiohttp-retry==2.8.3
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
Bootstrap-Flask==2.3.3
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.8
colorama==0.4.6
colorgram.py==1.2.0
distlib==0.3.7
dominate==2.9.1
filelock==3.12.2
Flask==3.0.0
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==23.0.0
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.1.2
packaging==24.2
pandas==2.2.3
pillow==10.4.0
platformdirs==3.9.1
propcache==0.2.0
pygame==2.6.1
PyJWT==2.10.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.31.0
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.25
twilio==9.3.7
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.0.4
virtualenv==20.29.2
virtualenvwrapper-win==1.2.7
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.0.1
yarl==1.18.0
